requirements.pip.txt
attrs==21.2.0
cffi==1.14.5
click==8.0.1
cryptography==3.4.7
decorator==4.4.2
et-xmlfile==1.1.0
flake8==3.9.2
hat-aio==0.4.1.dev20210601
hat-asn1==0.4.1.dev20210601
hat-drivers==0.4.1.dev20210601
hat-json==0.4.1.dev20210601
hat-peg==0.4.1.dev20210601
hat-util==0.4.1.dev20210601
jsonpatch==1.32
jsonpointer==2.1
jsonschema==3.2.0
llvmlite==0.36.0
mccabe==0.6.1
networkx==2.5.1
numba==0.53.1
numpy==1.21.0
openpyxl==3.0.7
packaging==20.9
pandapower==2.6.0
pandas==1.2.5
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
pyparsing==2.4.7
pyrsistent==0.17.3
pyserial==3.5
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
scipy==1.6.0
six==1.16.0
xlrd==2.0.1
XlsxWriter==1.4.3
